move Login component into Auth folder. add logout redirect uri.
