+++ src/components/Catalog/Catalog.tsx
@@ -1,28 +1,21 @@
-import React, { useEffect, useRef, useState } from "react";
-import { IonContent, IonPage, IonButton } from "@ionic/react";
+import React, { useState } from "react";
+import { IonContent, IonPage } from "@ionic/react";
 import { useHistory } from "react-router-dom";
 
 import { TabMenuLabels } from "../../constants/ui";
 import { getDate } from "../../utils/date";
 import { useCatalogQuery } from "../../data/useCatalogQuery";
-import { useAuth } from "../../store/AuthContext";
 
 import TerraLoader from "@nasa-terra/components/dist/react/loader";
-import TerraLogin from "@nasa-terra/components/dist/react/login";
-import TerraButton from "@nasa-terra/components/dist/react/button";
 import Banner from "../UI/Banner";
 import Variables from "./Variables";
 import InfoPanel from "../UI/InfoPanel";
-import Login from "./Login";
-
-const EDL_DOMAIN = "https://uat.urs.earthdata.nasa.gov";
+import Login from "../Auth/Login";
 
 const Catalog: React.FC = () => {
   const [variableId, setVariableId] = useState("");
   const history = useHistory();
   const { data: catalog, isLoading, isFetching } = useCatalogQuery();
-
-  const { user } = useAuth();
 
   const currentVariable = catalog?.find(
     (data) => data.dataFieldId === variableId,
@@ -86,14 +79,6 @@ const Catalog: React.FC = () => {
     <IonPage>
       <IonContent>
         <Banner>
-          {/* <IonButton
-            slot="end"
-            size="small"
-            onClick={() => setIsStorageOpen(true)}
-          >
-            <IonIcon aria-hidden="true" size="medium" icon={server} />
-          </IonButton> */}
-
           <Login />
         </Banner>
 
@@ -103,8 +88,6 @@ const Catalog: React.FC = () => {
           afterDismiss={afterInfoPanelDismiss}
         />
         <div className="ion-padding">
-          {/* TODO: check if user exists */}
-          <p>Welcome {user?.first_name}.</p>
           {isLoading && (
             <TerraLoader indeterminate variant="large"></TerraLoader>
           )}
